Formatting and dependencies update
